refactor(assets): 기술스택 아이콘을 Simple Icons 기반 SVG 로 통일
PNG 14 종을 Simple Icons / 공식 SVG 로 교체하면서 각 파일 fill 속성에
브랜드 hex 하드코딩. 스킬 카탈로그 확장을 위해 typescript / react_query /
claude_code / vite / vitest / zustand / pandacss 7 종을 assetManifest 에
신규 등록. assetManifest 와 StatusBar 의 PNG import 경로를 대응 SVG 로
전환.
