+++ src/shared/lib/assetManifest.ts
@@ -4,20 +4,23 @@ import folderFull from "@images/icons/folder_full.png";
 import imageDefault from "@images/icons/image_default.png";
 import documentDefault from "@images/icons/document_default.png";
 import chrome from "@images/icons/chrome.png";
-import react from "@images/icons/react.png";
-import javascript from "@images/icons/javascript.png";
-import redux from "@images/icons/redux.png";
-import recoil from "@images/icons/recoil.png";
-import nodejs from "@images/icons/nodejs.png";
-import html from "@images/icons/html.png";
-import css from "@images/icons/css.png";
-import styledcomponent from "@images/icons/styledcomponent.png";
-import github from "@images/icons/github.png";
-import vue from "@images/icons/vue.png";
-import python from "@images/icons/python.png";
-import bootstrap from "@images/icons/bootstrap.png";
-import tailwindCss from "@images/icons/tailwind-css.png";
-import vuetify from "@images/icons/vuetify.png";
+import react from "@images/icons/react.svg";
+import javascript from "@images/icons/javascript.svg";
+import typescript from "@images/icons/typescript.svg";
+import redux from "@images/icons/redux.svg";
+import recoil from "@images/icons/recoil.svg";
+import reactQuery from "@images/icons/reactquery.svg";
+import html from "@images/icons/html5.svg";
+import css from "@images/icons/css.svg";
+import styledcomponent from "@images/icons/styledcomponents.svg";
+import github from "@images/icons/github.svg";
+import vue from "@images/icons/vuedotjs.svg";
+import vuetify from "@images/icons/vuetify.svg";
+import claudeCode from "@images/icons/claude.svg";
+import vite from "@images/icons/vite.svg";
+import vitest from "@images/icons/vitest.svg";
+import pandacss from "@images/icons/pandacss.svg";
+import zustand from "@images/icons/zustand-plain.svg";
 import kdhProfile from "@images/김도현.jpg";
 import githubLine from "@images/icons/github_line.png";
 import blogLine from "@images/icons/blog_line.png";
@@ -40,18 +43,21 @@ const assetManifest: Record<string, string> = {
     chrome,
     react,
     javascript,
+    typescript,
     redux,
     recoil,
-    nodejs,
+    react_query: reactQuery,
     html,
     css,
     styledcomponent,
     github,
     vue,
-    python,
-    bootstrap,
-    tailwind_css: tailwindCss,
     vuetify,
+    claude_code: claudeCode,
+    vite,
+    vitest,
+    pandacss,
+    zustand,
     kdh_profile: kdhProfile,
     github_line: githubLine,
     blog_line: blogLine,
